blog-api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.1
pyjwt==2.1.0
python-dotenv==0.19.0
werkzeug==2.0.1
boto3==1.18.0
mangum==0.12.0